third_party/WebKit/LayoutTests/fast/dom/document-selectedStylesheetSet-use-counter-002.html

<!DOCTYPE html>
<title>document preferredStylesheetSet UseCounter</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
</style>
<script>
test(function() {
  const DocumentGetSelectedStylesheetSet = 2212; // From enums.xml
  const DocumentSetSelectedStylesheetSet = 2213; // From enums.xml
  assert_false(internals.isUseCounted(document, DocumentSetSelectedStylesheetSet));
  document.selectedStylesheetSet = '';
  assert_true(internals.isUseCounted(document, DocumentSetSelectedStylesheetSet));
  assert_false(internals.isUseCounted(document, DocumentGetSelectedStylesheetSet));
}, 'document set selectedStylesheetSet is use counted');
</script>
